Create Cogserver.html
main script that sits on the same server as the Cognos installation.  In an ideal environment this should simply be called from the application that wants to integrate the Cognos content.  Due to issues with CORS and environment security policy using an iFrame on the app server was required to avoid XSS warnings being generated.  This has been abandoned in favor of a Java/SOAP implementation.